+++ pkg/exchange/binance/singleton_test.go
@@ -1,0 +1,7 @@
+package binance
+
+import "testing"
+
+func TestExit(t *testing.T) {
+	// TODO: 完成这个测试
+}